+++ CHANGELOG.md
@@ -20,7 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [x] leading zero's in ip4/6 prefix lengths is actually a syntax error
 - [x] empty macro-string in an *unknown* modifier is actually legal
 - [x] %{t} now expands to timestamp (UNIX epoch time)
-- [ ] unknown modifiers also cannot use c,r,t-macros
+- [x] unknown modifiers also cannot use c,r,t-macros
 
 ### changed